Move entity and triple stores to correct modules
