+++ apps/web/modules/action/index.ts
@@ -1,3 +1,4 @@
 export { ActionsStore } from './actions-store';
 export { ActionsStoreProvider, useActionsStoreContext } from './actions-store-provider';
+export { useActionsStore } from './use-actions-store';
 export * as Action from './action';
