+++ apps/web/modules/action/use-actions-store.tsx
@@ -1,7 +1,7 @@
 import { useSelector } from '@legendapp/state/react';
 import { useActionsStoreContext } from './actions-store-provider';
 
-export function useActions(spaceId: string) {
+export function useActionsStore(spaceId: string) {
   const { actions$, publish } = useActionsStoreContext();
   const actions = useSelector(actions$);
