+++ apps/web/modules/triple/index.ts
@@ -1,1 +1,3 @@
 export * as Triple from './triple';
+export { TripleStoreProvider, useTripleStoreContext } from './triple-store-provider';
+export { TripleStore } from './triple-store';
