+++ apps/web/pages/space/[id].tsx
@@ -4,7 +4,7 @@ import { Triples } from '~/modules/components/triples';
 import { SYSTEM_IDS } from '~/modules/constants';
 import { Network } from '~/modules/services/network';
 import { StorageClient } from '~/modules/services/storage';
-import { TripleStoreProvider } from '~/modules/stores/triple-store-provider';
+import { TripleStoreProvider } from '~/modules/triple/triple-store-provider';
 import { Triple } from '~/modules/types';
 import { Params } from '~/modules/params';
 import { DEFAULT_PAGE_SIZE } from '~/modules/stores/triple-store';
